Implementation complete
Now game has a struct with all the NPCS, whenever an NPC is added from the .dat, each one is placed on the space with the corresponding id, because space now has the pointers to the NPCs stored in that space.

Also added some functions to move, add and remove NPCs from the space.

IMPORTANT: Do. Not. Try. tk O470